refactor: decompoe componente Relatorios em View e subcomponentes
- Substitui o componente monolítico Relatorios.vue por RelatoriosView.vue
- Extrai filtros para components/relatorios/RelatoriosFiltros.vue
- Extrai visualizações específicas: RelatoriosFarmacia e RelatoriosFimPlantao
- Implementa RelatoriosTabela para visualizações genéricas
- Atualiza rotas para utilizar a nova View

Fixes #11
