+++ frontend/src/router/index.ts
@@ -7,7 +7,7 @@ import PacientesView from '@/views/PacientesView.vue'
 import AgendaView from '@/views/AgendaView.vue'
 import Agendamento from '@/components/Agendamento.vue'
 import Farmacia from '@/components/Farmacia.vue'
-import Relatorios from '@/components/Relatorios.vue'
+import RelatoriosView from '@/views/RelatoriosView.vue'
 import ProtocolosView from '@/views/ProtocolosView.vue'
 import Ajustes from '@/components/Ajustes.vue'
 import PrescricaoView from '@/views/PrescricaoView.vue'
@@ -22,7 +22,7 @@ const router = createRouter({
         {path: 'agenda', name: 'Agenda', component: AgendaView},
         {path: 'agendamento', name: 'Agendamento', component: Agendamento},
         {path: 'farmacia', name: 'Farmacia', component: Farmacia},
-        {path: 'relatorios', name: 'Relatorios', component: Relatorios},
+        {path: 'relatorios', name: 'Relatorios', component: RelatoriosView},
         {path: 'protocolos', name: 'Protocolos', component: ProtocolosView},
         {path: 'ajustes', name: 'Ajustes', component: Ajustes},
         {path: 'prescricao', name: 'Prescricao', component: PrescricaoView}
